nightlife/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
channels==4.0.0
charset-normalizer==3.3.2
Django==5.0.3
django-braces==1.15.0
django-cors-headers==4.3.1
django-cron==0.6.0
django-leaflet==0.29.0
django-model-utils==4.4.0
django-star-ratings==0.9.2
djangorestframework==3.14.0
googlemaps==4.10.0
idna==3.6
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
swapper==1.3.0
tzdata==2024.1
urllib3==2.2.1
